requirements.txt
jinja2==2.7.3
Flask==0.10.1
requests==2.2.1
redis==2.9.1
rethinkdb==2.0.0-1
nose==1.3.0
gevent==1.0.1
gunicorn==19.0.0
simplejson==3.6.2
uWSGI==2.0.7
psycopg2
Flask-SQLAlchemy
Flask-Migrate
